+++ tcex/tcex_batch_v2.py
@@ -48,7 +48,8 @@ def custom_indicator_class_factory(indicator_type, base_class, class_dict, value
 class TcExBatch(object):
     """ThreatConnect Batch Import Module"""
 
-    def __init__(self, tcex, owner, action=None, attribute_write_type=None, halt_on_error=True):
+    def __init__(self, tcex, owner, action=None, attribute_write_type=None, halt_on_error=True,
+                 playbook_triggers_enabled=None):
         """Initialize Class Properties.
 
         Args:
@@ -65,6 +66,9 @@ class TcExBatch(object):
         self._batch_max_chunk = 5000
         self._halt_on_error = halt_on_error
         self._owner = owner
+        self._playbook_triggers_enabled = playbook_triggers_enabled
+
+        # shelf settings
         self._group_shelf_fqfn = None
         self._indicator_shelf_fqfn = None
 
@@ -535,8 +539,9 @@ class TcExBatch(object):
                     'type': group_data.get('type')
                 }
         else:
+            GROUPS_STRINGS_WITH_FILE_CONTENTS = ['Document', 'Report']
             # process file content
-            if group_data.data.get('type') in ['Document', 'Report']:
+            if group_data.data.get('type') in GROUPS_STRINGS_WITH_FILE_CONTENTS:
                 self._files[group_data.data.get('xid')] = group_data.file_data
             group_data = group_data.data
         return group_data
@@ -979,7 +984,7 @@ class TcExBatch(object):
         if self._poll_interval is None and self._batch_data_count is not None:
             # calculate poll_interval base off the number of entries in the batch data
             # with a minimum value of 5 seconds.
-            self._poll_interval = max(math.ceil(self._batch_data_count / 250), 5)
+            self._poll_interval = max(math.ceil(self._batch_data_count / 300), 5)
         elif self._poll_interval is None:
             # if not able to calculate poll_interval default to 15 seconds
             self._poll_interval = 15
@@ -1027,7 +1032,8 @@ class TcExBatch(object):
 
             if data.get('data', {}).get('batchStatus', {}).get('status') == 'Completed':
                 # store last 5 poll times to use in calculating average poll time
-                self._poll_interval_times = self._poll_interval_times[-4:] + [(poll_time_total * .7)]
+                modifier = poll_time_total * .7
+                self._poll_interval_times = self._poll_interval_times[-4:] + [modifier]
 
                 weights = [1]
                 poll_interval_time_weighted_sum = 0
@@ -1189,7 +1195,7 @@ class TcExBatch(object):
     @property
     def settings(self):
         """Return batch job settings."""
-        return {
+        _settings = {
             'action': self._action,
             # not supported in v2 batch
             # 'attributeWriteType': self._attribute_write_type,
@@ -1198,6 +1204,9 @@ class TcExBatch(object):
             'owner': self._owner,
             'version': 'V2'
         }
+        if self._playbook_triggers_enabled is not None:
+            _settings['playbookTriggersEnabled'] = str(self._playbook_triggers_enabled).lower()
+        return _settings
 
     def signature(self, name, file_name, file_type, file_text, xid=True):
         """Add Signature data to Batch object.
@@ -1254,6 +1263,7 @@ class TcExBatch(object):
             halt_on_error).get('data', {}).get('batchStatus', {})
         batch_id = batch_data.get('id')
         if batch_id is not None:
+            self.tcex.log.info('Batch ID: {}'.format(batch_id))
             # job hit queue
             if poll:
                 # poll for status
@@ -1632,6 +1642,7 @@ class Group(object):
         }
         self._attributes = []
         self._labels = []
+        self._file_content = None
         self._tags = []
         # processed
         self._processed = False
@@ -1649,6 +1660,21 @@ class Group(object):
             xid = str(uuid.uuid4())
 
         return xid
+
+    def add_file(self, filename, file_content):
+        """Add a file for Document and Report types.
+
+        Example::
+
+            document = tcex.batch.group('Document', 'My Document')
+            document.add_file('my_file.txt', 'my contents')
+
+        Args:
+            filename (str): The name of the file.
+            file_content (bytes|method|str): The contents of the file or callback to get contents.
+        """
+        self._group_data['fileName'] = filename
+        self._file_content = file_content
 
     def add_key_value(self, key, value):
         """Add custom field to Group object.
@@ -1735,6 +1761,26 @@ class Group(object):
                 if tag.valid:
                     self._group_data['tag'].append(tag.data)
         return self._group_data
+
+    @property
+    def date_added(self):
+        """Return Group dateAdded."""
+        return self._group_data.get('dateAdded')
+
+    @date_added.setter
+    def date_added(self, date_added):
+        """Set Indicator dateAdded."""
+        self._group_data['dateAdded'] = self._utils.format_datetime(
+            date_added, date_format='%Y-%m-%dT%H:%M:%SZ')
+
+    @property
+    def file_data(self):
+        """Return Group file (only supported for Document and Report)."""
+        return {
+            'fileContent': self._file_content,
+            'fileName': self._group_data.get('fileName'),
+            'type': self._group_data.get('type')
+        }
 
     @property
     def name(self):
@@ -1876,11 +1922,7 @@ class Document(Group):
         super(Document, self).__init__('Document', name, xid)
         self._group_data['fileName'] = file_name
         # file data/content to upload
-        self._file_data = {
-            'fileContent': file_content,
-            'fileName': file_name,
-            'type': self._group_data.get('type')
-        }
+        self._file_content = file_content
         if malware:
             self._group_data['malware'] = malware
         if password is not None:
@@ -1889,17 +1931,21 @@ class Document(Group):
     @property
     def file_content(self):
         """Return Group files."""
-        return self._file_data.get('fileContent')
+        return self._file_content
 
     @file_content.setter
     def file_content(self, file_content):
         """Set Document or Report file data."""
-        self._file_data['fileContent'] = file_content
+        self._file_content = file_content
 
     @property
     def file_data(self):
         """Return Group files."""
-        return self._file_data
+        return {
+            'fileContent': self._file_content,
+            'fileName': self._group_data.get('fileName'),
+            'type': self._group_data.get('type')
+        }
 
     @property
     def malware(self):
@@ -2124,11 +2170,7 @@ class Report(Group):
         super(Report, self).__init__('Report', name, xid)
         self._group_data['fileName'] = file_name
         # file data/content to upload
-        self._file_data = {
-            'fileContent': file_content,
-            'fileName': file_name,
-            'type': self._group_data.get('type')
-        }
+        self._file_content = file_content
         if publish_date is not None:
             self._group_data['publishDate'] = self._utils.format_datetime(
                 publish_date, date_format='%Y-%m-%dT%H:%M:%SZ')
@@ -2136,17 +2178,21 @@ class Report(Group):
     @property
     def file_content(self):
         """Return Group files."""
-        return self._file_data.get('fileContent')
+        return self._file_content
 
     @file_content.setter
     def file_content(self, file_content):
         """Set Document or Report file data."""
-        self._file_data['fileContent'] = file_content
+        self._file_content = file_content
 
     @property
     def file_data(self):
         """Return Group files."""
-        return self._file_data
+        return {
+            'fileContent': self._file_content,
+            'fileName': self._group_data.get('fileName'),
+            'type': self._group_data.get('type')
+        }
 
     @property
     def publish_date(self):
@@ -2396,6 +2442,28 @@ class Indicator(object):
                 if tag.valid:
                     self._indicator_data['tag'].append(tag.data)
         return self._indicator_data
+
+    @property
+    def date_added(self):
+        """Return Indicator dateAdded."""
+        return self._indicator_data.get('dateAdded')
+
+    @date_added.setter
+    def date_added(self, date_added):
+        """Set Indicator dateAdded."""
+        self._indicator_data['dateAdded'] = self._utils.format_datetime(
+            date_added, date_format='%Y-%m-%dT%H:%M:%SZ')
+
+    @property
+    def last_modified(self):
+        """Return Indicator lastModified."""
+        return self._indicator_data.get('lastModified')
+
+    @last_modified.setter
+    def last_modified(self, last_modified):
+        """Set Indicator lastModified."""
+        self._indicator_data['lastModified'] = self._utils.format_datetime(
+            last_modified, date_format='%Y-%m-%dT%H:%M:%SZ')
 
     def occurrence(self, file_name=None, path=None, date=None):
         """Add a file Occurrence.