+++ docs/src/release_notes.rst
@@ -6,9 +6,24 @@ Release Notes
 0.8.x
 =====
 
-0.8.19 (working)
+0.8.21
+------
++ Added PDF method to resource module for supported group types.
++ Added task_id method for Task class.
++ Added date_added property to Indicator and Groups objects.
++ Added last_modified property to Indicator objects.
++ Updated tcrun for handling Binary/BinaryArray validation.
+
+0.8.20
+------
++ Fixed deletion in the batch module for TC instances < 5.7.
+
+0.8.19
 ------
 + Removed app.lock logic.
++ Updated file_content logic for Documents and Reports.
++ Added ``add_file()`` method for batch Group objects.
++ Added playbook_triggers_enabled parameter to batch module (requires ThreatConnect 5.7).
 
 0.8.18
 ------